# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nestling API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session cookie issued"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many attempts from this IP"
          }
        }
      }
    },
...
    "/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Family and first parent created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredResponse"
                }
              }
            }
          },
          "403": {
            "description": "Registration disabled"
          },
          "409": {
            "description": "Email already registered"
          },
          "422": {
            "description": "Password too short"
          }
        }
      }
    },
    "/v1/children": {
      "get": {
        "tags": [
          "children"
        ],
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Children in this family",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChildResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      },
      "post": {
        "tags": [
          "children"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewChild"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Child created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Empty display name"
          }
        }
      }
    },
    "/v1/children/{id}": {
      "delete": {
        "tags": [
          "children"
        ],
        "operationId": "soft_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Child id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Soft deleted"
          },
          "404": {
            "description": "No such child in this family"
          }
        }
      }
    },
...
    "/v1/children/{id}/enrollments": {
      "post": {
        "tags": [
          "devices"
        ],
        "operationId": "mint_enrollment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Child id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "One-time enrollment code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such child in this family"
          }
        }
      }
    },
...
    "/v1/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "operationId": "list_devices",
        "responses": {
          "200": {
            "description": "Devices in this family",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/v1/devices/{id}": {
      "delete": {
        "tags": [
          "devices"
        ],
        "operationId": "revoke_device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Token revoked"
          },
          "404": {
            "description": "No such device in this family"
          }
        }
      }
    },
...
    "/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "The signed-in parent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChildResponse": {
        "type": "object",
        "required": [
          "id",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Credentials": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "DeviceResponse": {
        "type": "object",
        "required": [
          "id",
          "child_id",
          "platform",
          "model",
          "label",
          "revoked",
          "stale"
        ],
        "properties": {
          "child_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "model": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "revoked": {
            "type": "boolean"
          },
          "stale": {
            "type": "boolean"
          }
        }
      },
...
      "EnrollmentResponse": {
        "type": "object",
        "required": [
          "code",
          "expires_at",
          "qr_payload"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "qr_payload": {
            "type": "string"
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "family_id"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "NewChild": {
        "type": "object",
        "required": [
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          }
        }
      },
...
      "RegisteredResponse": {
        "type": "object",
        "required": [
          "family_id"
        ],
        "properties": {
          "family_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}